Expose default implementation through package object
