+++ src/main/scala/scalaz/distributed/Client.scala
@@ -2,7 +2,11 @@ package scalaz.distributed
 
 import java.net.InetAddress
 
+import scalaz.zio.IO
+
 trait Client {
+  type Distributed[A] = IO[Error, A]
+
   type Type[A]
 
   trait Metadata {
@@ -21,21 +25,4 @@ trait Client {
   def connect(member: Member, seed: Set[InetAddress]): Distributed[Protocol]
 
   def members(cb: Membership => Distributed[Boolean]): Distributed[Unit]
-}
-
-object Client {
-
-  def default: Client =
-    new Client {
-      type Type[A] = SupportedType[A]
-
-      override def key[K: Type, V: Type](k: K): Path[Map[K, V], V] = Path.Key[K, V](k)
-
-      override def compose[A: Type, B: Type, C: Type](x: Path[A, B], y: Path[B, C]): Path[A, C] =
-        x >>> y
-
-      override def connect(member: Member, seed: Set[InetAddress]): Distributed[Protocol] = ???
-
-      override def members(cb: Membership => Distributed[Boolean]): Distributed[Unit] = ???
-    }
 }
